renaming snippets files for better cohesion and comprehension
